% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ask.R
\name{Ask}
\alias{Ask}
\title{Prompts for user input.}
\usage{
Ask(...)
}
\description{
\code{Ask} asks for user input through the web interface and returns the
given values in a list.
}
\details{
The following types of inputs are allowed:
}

