% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{Bernoulli}
\alias{Bernoulli}
\title{Bernoulli Sampling}
\usage{
Bernoulli(data, p = 0.5, rng = "mt19937_64")
}
\arguments{
\item{data}{A \code{\link{waypoint}}.}

\item{p}{The probability that a tuple passes the filter.}

\item{rng}{A length-one character vector specifying the random number
generator used. It should correspond to an engine in the C++ header
<random>.}
}
\value{
A \code{\link{waypoint}} that has been filtered randomly.
}
\description{
The data is filtered randomly using a Bernoulli trial for each tuple.
}
\section{author}{

Jon Claus, \email{jonterainsights@gmail.com}, Tera Insights, LLC.
}
\references{
href{http://en.cppreference.com/w/cpp/numeric/random}{C++ reference} for a
list of random number generators.
href{http://en.wikipedia.org/wiki/Bernoulli_sampling}{Wikipedia} for
information about Bernoulli sampling.
}

