% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count.R
\name{Count}
\alias{Count}
\title{Number of Rows in a Waypoint.}
\usage{
Count(data, outputs = count)
}
\arguments{
\item{data}{A \code{\link{waypoint}} object.}

\item{outputs}{The column name of the result.}
}
\value{
A \code{\link{waypoint}} containing a single row and column whose
  name is give by \code{output}.
}
\description{
Counts the total number of tuples in the given waypoints.
}
\author{
Jon Claus, <jonterainsights@gmail.com>, Tera Insights, LLC.
}
\seealso{
\code{\link{BloomFilter}} for a similar GLA.
}

