% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.R
\name{Create}
\alias{Create}
\alias{Delete}
\title{Create and Delete Relations.}
\usage{
Create(name, ...)

Delete(name)
}
\arguments{
\item{...}{Named arguments used to specify the schema of the relation being
created. Each argument should be in the form of \code{name = type}, where
\code{name} is a valid name for an attribute and type specifies a
\code{\link{type}}.}

\item{relation}{Usually, a name or character string specifying the relation
  to load. A character string (enclosed in explicit single or double quotes)
  is always taken as the relation name.

  If the value of \code{relation} is a length-one character vector the name
  of the relation is taken to be the value of the only element. Otherwise,
  \code{relation} must be a name or character string.}
}
\value{
An \code{\link{invisible}} NULL value.
}
\description{
Functions used to create and delete relations.
}
\details{
These functions are used to create new relations. These functions only affect
the relational schemas visible to the user; they have no actual effect on the
hard disk storage. In order to actually write data to an existing relation,
\code{\link{Store}} should be used. When \code{Delete} is called, that
relation is no longer available to the user and the space is used on the hard
disc is freed, but the stored data is not immediately overwritten.
}
\author{
Jon Claus, <jonterainsights@gmail.com>, Tera Insights, LLC.
}

