% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distinct.R
\name{Distinct}
\alias{Distinct}
\title{Remove Duplication Combinations}
\usage{
Distinct(data, inputs, outputs)
}
\arguments{
\item{inputs}{A named list of expressions, with the names being used as the
  corresponding outputs. These expressions are outputted in addition to those
  used to specify the extremities.

  If no name is given and the corresponding  expression is simply an
  attribute, then said attribute is used as the name. Otherwise an error is
  thrown, as there is no reason to include an extra input if corresponding
  output column cannot be referenced later.}

\item{outputs}{The usual way to specify the outputs. If both this and names
for the inputs are given, an error is thrown.}
}
\value{
A \code{\link{waypoint}}.
}
\description{
Filters out repeated combination of inputs.
}
\details{
This GLA returns the distinct combinations of the given inputs using a full
hashing of the distinct combinations. As such, it requires \eqn{O(k)} space,
where \eqn{k} is the number of distinct combinations. The run time is
\eqn{O(n + k)}, where \eqn{n} is the number of rows in \code{data}. The
second term is a result of having to merge hashes between different states.
Having a large number of distinct values leads to significant slowdown
because of this.
}
\author{
Jon Claus, <jonterainsights@gmail.com>, Tera Insights, LLC.
}

