% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtBase.R
\name{expressions}
\alias{expressions}
\title{Expressions}
\description{
Expressions are used as inputs for various waypoints.
}
\details{
An expression specified in the following manner:

\begin{description}
\item[Test] this is a test.
\item[foo] blah
\end{description}
}
\note{
For the most part, the expression is not evaluated at all. Instead, it
is parsed and its abstract syntax tree is visited and translated.
}
\examples{
## Basic expression
(att1 + att2) * att3

## Use of .()
a <- 1
(att1 + .(a)) * att3
}
\author{
Jon Claus, <jonterainsights@gmail.com>, Tera Insights LLC
}

