% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/groupby.R
\name{GroupBy}
\alias{GroupBy}
\title{Compute Aggregates of Data Subsets}
\usage{
GroupBy(data, group, ..., fragment.size = 2e+06, init.size = 1024,
  use.mct = TRUE, debug = 0, states = list())
}
\arguments{
\item{data}{A \code{\link{waypoint}}.}

\item{group}{A named list of expressions, with the names being used as the
  corresponding outputs. These expressions are outputted in addition the
  results of the inner GLAs.

  If no name is given and the corresponding  expression is simply an
  attribute, then said attribute is used as the name. Otherwise, the column
  for that expression is hidden from the user.}

\item{fragment.size}{The number of tuples returned per fragment. This should
only be changed from its default value by experienced users.}

\item{init.size}{The number of groups that space is initially allocated for.}

\item{use.mct}{Should the MCT hash function be used.}

\item{debug}{Should debugging information be printed to standard output.}

\item{states}{Additional states to pass through.}

\item{\dots}{Specification of the inner GLAs. See \sQuote{details} for more
information.}
}
\value{
A \code{\link{waypoint}}.
}
\description{
Splits the data into subsets based on grouping expressions, performs the
given aggregates per subset, and returns the results in a convenient form.
}
\details{
The inner GLAs should be specified as a list of calls to other aggregate
functions, such as Sum or Mean. In each of these calls, the \code{data}
argument should be omitted, as it is inferred to be the \code{data} passed to
\code{GroupBy}. Additionally, each argument specifying an inner GLA may be
named. If so, that name is taken to be the output of the corresponding GLA.
This is purely a stylistic shortcut and the normal method of specifying the
outputs can still be used instead.

The outputs of these inner GLAs should avoid name clashing both with each
other and those for the grouping expressions.

In the case that one inner GLA produces multiple rows and the rest produce a
single row, each of the single row outputs are repeated accordingly.

If more than one inner GLA produces multiple rows, an error is thrown.

The output for each group is then concatenated and returned.
}
\author{
Jon Claus, <jonterainsights@gmail.com>, Tera Insights, LLC.
}

