% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.R
\name{Load}
\alias{Load}
\alias{Read}
\title{Load a relation.}
\usage{
Load(relation)

Read(relation)
}
\arguments{
\item{relation}{Usually, a name or character string specifying the relation
  to load. A character string (enclosed in explicit single or double quotes)
  is always taken as the relation name.

  If the value of \code{relation} is a length-one character vector the name
  of the relation is taken to be the value of the only element. Otherwise,
  \code{relation} must be a name or character string.}
}
\value{
A \code{waypoint} object whose schema is determined by the
  relation being loaded.
}
\description{
\code{Load} loads a binary relation from the disc.
}
\details{
An error is thrown if \code{relation} does not specify a relation that exists
and can be read by the user.

\code{Read} is simply an alias for \code{Load} that exists for compatibility.
}

