% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{[.data}
\alias{[.data}
\title{Basic Filtering of Waypoints}
\usage{
\method{[}{data}(data, condition)
}
\arguments{
\item{data}{A \code{\link{waypoint}}.}

\item{condition}{A boolean valued \code{\link{expression}}.}
}
\value{
A \code{\link{waypoint}} with a subset of the tuples in \code{data}.
}
\description{
Filter a waypoint based on a boolean expression.
}
\details{
\code{condition} is evaluated for each tuple in \code{data} independently.
Only those tuples for which \code{condition} evaluates to TRUE are passed
through the filter.
}
\author{
Jon Claus, <jonterainsights@gmail.com>, Tera Insights, LLC.
}

